apps/batch-worker/pyproject.toml
[project]
name = "batch-worker"
version = "0.1.0"
description = "PDF一括解析バッチ処理システムのワーカーコンポーネント"
requires-python = ">=3.12"
dependencies = [
    "redis>=5.0.0",
    "google-cloud-pubsub>=2.25.0",
    "google-cloud-storage>=2.18.0",
    "pydantic-settings>=2.6.0",
    "loguru>=0.7.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.ruff]
line-length = 100
target-version = "py312"

[tool.ruff.lint]
select = [
    "E",   # pycodestyle errors
    "W",   # pycodestyle warnings
    "F",   # pyflakes
    "I",   # isort
    "B",   # flake8-bugbear
    "C4",  # flake8-comprehensions
    "UP",  # pyupgrade
]
ignore = []

[tool.ruff.lint.isort]
known-first-party = ["config", "storage", "processor"]

[tool.mypy]
python_version = "3.12"
strict = true
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
